Reload (rather than restart) pgbouncer on configuration change
Without this change, PgBouncer reloads (using systemctl reload
pgbouncer@xxx) use pgbouncer's -R option.  This is a problem because it
causes an interruption to availability of pgbouncer (and hence the
database).

In addition to updating the reload command in the systemd unit file, it also
configures Puppet's pgbouncer service management to use that updated reload
command when configuration is changed.